always apply classTableWrapper option
previously, the classTableWrapper was only applied in case a pager was added to the table (table has at least smallest page size).
The scrolling template thus only worked for paging tables and not for wide but too short tables.
Changed this (and changed the scrolling template and its CSS in order to be applicable if the outer wrappers are missing, which still are only applied together with a pager).