Aria-Labels, modified arrowicons template
NEW: Introduced Aria-Labels to the pager buttons for better accessibility
NEW: option buttonLabelClass
NEW: Default CSS now contains .flipLeft and .rotateLeft classes which both can be used as values for the new buttonLabelClass option if the buttonLabels for the left buttons (first/prev) are in fact right-pointing.
UPD: Updated the template "triangleIcons" to use the same right-pointing arrow icon for all buttons and flip the left buttons using the aforementioned options and class. Old template conserved as "triangleIconsOld".